FlipGraphs.jl Documentation

Welcome to the documentation of the FlipGraphs.jl package. This package was created as part of my master's thesis on triangulations of closed surfaces and the flip graphs thereof.

The aim of the package is to model triangulations on any closed surface and compute their respective flip graphs.

FlipGraphs — Module

FlipGraphs

A package for triangulations of points and their respective flip graphs.

So far, this package includes:
- triangulations of convex polygons / triangulations of points situated on the border of a disc and their flip graphs (with labeled and unlabeled points).
- triangulations of points on a closed surface and their respective modular flip graphs (with labeled and unlabeled points).

Table of contents

Installation

Quick Start

Triangulated Convex Polygon

Δ-Complex / Triangulation of closed surface

Modular FlipGraph

Triangulations of Convex Polygons

Structures

Constructors

Graph Methods

Flip an edge

Triangulations of Closed Surfaces

DeltaComplex

Classifying the triangulation

Non-orientable closed surfaces

Flipping

Modular Flip Graphs of Closed Orientable Surfaces

Structures

Constructors

Graph methods

Comparing Triangulations

Canonical labeling of DeltaComplexes

Exporting

General utilities