FlipGraph

Contrary to flipgraphs of planar triangulations like that of a convex polygon, the flipgraph of a closed surface is generally infinitely large. Therefore, it is impossible to construct the whole flipgraph. Not only are they infinitely in size, but they also "grow" very fast. It is therefore more interesting and achievable, to look at the modular flip graphs. These are graphs whose vertices are isotopy classes of triangulations.

FlipGraphs.FlipGraph — Type

struct FlipGraph <: AbstractGraph{Int}

A Graph representing the flipgraph of a Δ-Complex.

Vertices are different triangulations of the same surface.
Two vertices are linked by an edge, if the respective graphs differ only by a single flip.

FlipGraphs.FGVertex — Type

struct FGVertex

A vertex in a flip graph.

An FGVertex is composed of a representant(DeltaComplex) of the isotopy class of that vertex. The representant has been relabeld with one of the canonical labelings obtained by McKay's Algorithm. Addidtionally, the FGVertex contains the number of labelings that are output by the respective McKay's Algorithms.

Graphs.nv — Method

nv(G::FlipGraph) -> Int

Return the number of vertices in G.

Graphs.ne — Method

ne(G::FlipGraph) -> Int

Return the number of edges in G.

Graphs.vertices — Method

vertices(G::FlipGraph) -> Vector{DeltaComplex}

Return a list of all the vertices that have been constructed in G.

Graphs.edges — Method

edges(G::FlipGraph) ::Vector{Edge}

Construct an array containing all the edges in G.

Graphs.has_vertex — Method

has_vertex(G::FlipGraph, v::Integer) -> Bool

Return true if v is a valid index of a vertex in G.

Graphs.has_edge — Method

has_edge(G::FlipGraph, e::Edge) -> Bool

Graphs.has_edge — Method

has_edge(G::FlipGraph, s::Integer, d::Integer) -> Bool

Graphs.neighbors — Method

neighbors(G::FlipGraph, v::Integer) -> Vector{Int}

Return a list of the indices of all the neighboring vertices of the v-th vertex.

Graphs.diameter — Method

diameter(G::FlipGraph)

Return the diameter of the portion of the flipgraph that has been constructed.

Comparing Triangulations

One problem of deciding, if two triangulations are equivalent, is that the naming of the vertices, edges and points is completely arbitrary. In the act of flipping, DualEdges and TriFaces "move around". It is therefore possible, to obtain two HoleyDeltaComplexs representing the same triangulation of a surface but with different labeled TriFaces and DualEdges. Checking every possible permutation of TriFace and DualEdge labelings is not an option, as the number of possibilities would blow up immeadiately.

What we do instead, is trying to find a canonical labeling. This Module uses a version of McKay's canonical graph labeling algorithm[1] to try and determin a unique labeling based on the relationship to other vertices and edges. In general, it is not always possible to determin a unique labeling.However, with this method, we can reduce the number of labelings to a manageable number.

The mcKay Methods each return a permutation vector p which can be interpreted as Cauchy's one-line notation for permutations. For example p = [3,5,1,2,6,4] would correspond to the following permutation:

\[\centering σ = \begin{pmatrix} 1 & 2 & 3 & 4 & 5 & 6\\ 3 & 5 & 1 & 2 & 6 & 4 \end{pmatrix} = (1 3)(2 5 6 4)\]

FlipGraphs.mcKay_points — Method

mcKay_points(D::DeltaComplex; only_one::Bool=false)::Vector{Vector{Int}}

Apply a version of McKay's canonical graph labeling algorithm in order to determine all possible permutations of the points which give a canonical isomorphism class representant.

Return a vector of permutation vectors p such that Point 1 becomes Point p[1], Point 2 becomes Point p[2],... If only_one=true, the algorithm stops after finding one valid permutation.

FlipGraphs.mcKay_vertices — Method

mcKay_vertices(D::DeltaComplex)::Vector{Vector{Int}}

Apply a version of McKay's canonical graph labeling algorithm in order to determine all possible permutations of the triFaces which give a canonical isomorphism class representant.

Return a vector of permutation vectors p such that TriFace 1 becomes TriFace p[1], TriFace 2 becomes TriFace p[2],...
If only_one=true, the algorithm stops after finding one valid permutation.

FlipGraphs.mcKay_edges — Method

mcKay_edges(D::DeltaComplex; only_one::Bool=false)::Vector{Vector{Int}}

Apply McKay's canonical graph labeling algorithm in order to determine all possible permutations of the triFaces which give a canonical isomorphism class representant.

Return a vector of permutation vectors p such that DualEdge 1 becomes DualEdge p[1], DualEdge 2 becomes DualEdge p[2],...
If only_one=true, the algorithm stops after finding one valid permutation.

FlipGraphs.is_isomorphic — Method

is_isomorphic(D1::DeltaComplex, D2::DeltaComplex; kwargs..) -> Bool

Return true if D1 is isomorph to D2 up to a renaming of the vertices, edges and if fix_points=false also points.

Arguments

fix_points::Bool=true : If fix_points is set to false, then the isomorphism also includes a renaming of the points.

FlipGraphs.is_isomorphic — Method

is_isomorphic(candidate::FGVertexCandidate, fgv::FGVertex; kwargs...) -> Bool

Return true if candidate is in the isotopy class of fgv.

Arguments

fix_points::Bool=true : If is set to false, then the isomorphism would also allow a relabeling of the points.

Constructing the FlipGraph

FlipGraphs.flipgraph_modular — Function

flipgraph_modular(g::Integer; kwargs..)

Construct the FlipGraph for the DeltaComplex D.

Arguments

fix_points::Bool=true : If is set to false, then the isomorphism also includes a renaming of the points.

flipgraph_modular(D::DeltaComplex; kwargs..)

Construct the FlipGraph for the DeltaComplex D.

Arguments

fix_points::Bool=true : If is set to false, then the isomorphism also includes a renaming of the points.

depth::Integer = ∞ : Determines the depth to which the flip grap should be constructed. i.e. up to which distance from D.

1Hartke, S.G., & Radcliffe, A.J. (2008). McKay ’ s Canonical Graph Labeling Algorithm.