Flip Graphs of Convex Polygons

Flip Graphs, are obtained, by considering triangulations on a fixed set of points as vertices, and connect two vertices, if it is possible to get from one triangulation to the other by flipping a single edge.

FlipGraphs.FlipGraphPlanar — Type

struct FlipGraphPlanar <: AbstractGraph{Int32}

A Graph representing the FlipGraph of a convex polygon.

Vertices are different triangulations of the same convex polygon.

Two vertices are linked by an edge, if the respective graphs differ only by a single flip.

FlipGraphPlanar implements the AbstractGraph interface from Graphs.jl. It is therefore possible to use it with other packages that work with Graphs.jl. This is very helpfull for plotting the graph.

Constructors

FlipGraphs.flipgraph — Method

flipgraph(g::TriangulatedPolygon , modular=false)

Construct the FlipGraph for the TriangulatedPolygon g.

If modular is true, then vertices of the FlipGraph are the classes of isomorphisms up to renaming the vertices. Each class is represented by one of its elements.
If modular is false, then each vertex is a different triangulation of the initial graph g.
By default, modular is set to false.

FlipGraphs.flipgraph_planar — Function

flipgraph_planar(n::Integer, modular=false)

Construct the FlipGraph of a convex n-gon.

If modular=true, the FlipGraph is reduced to its modular form.

Examples

julia> flipgraph_planar(6, false)
FlipGraphPlanar with 14 vertices and 21 edges

Graph methods

Graphs.nv — Method

nv(G::FlipGraphPlanar)

Return the number of vertices in G.

Graphs.ne — Method

ne(G::FlipGraphPlanar)

Return the number of edges in G.

Graphs.vertices — Method

vertices(G::FlipGraphPlanar)

Return the List of all vertices in G.

Graphs.edges — Method

edges(G::FlipGraphPlanar) ::Vector{Edge}

Construct an array containing all the edges in G.

Graphs.has_vertex — Method

has_vertex(G::FlipGraphPlanar, v)

Return true if v is a valid index of a vertex in G.

Graphs.has_edge — Method

has_edge(G::FlipGraphPlanar, s, d)

Return true if there is an edge between s and d in G.

Graphs.has_edge — Method

has_edge(G::FlipGraphPlanar, e::Edge)

Return true if e is an edge in G.

Graphs.neighbors — Method

neighbors(G::FlipGraphPlanar, v::Integer) -> Vector{Int32}

Return a list of all the indices of vertices in G, that are adjacent to v.

Graphs.diameter — Method

diameter(G::FlipGraphPlanar)

Compute the diameter of G.