HoleyDeltaComplex

The DeltaComplex structure is usefull for many applications. However, it has a slight issue that two different triangulations of the same surface, may result in the same DeltaComplex. In other words, a DeltaComplex misses some information about the underlying triangulation and can therefore not be used if one would, for example, like to construct the flipgraph (or rather a portion of latter) of the triangulations on a given surface.

The HoleyDeltaComplex structure solves this issue, by keeping track of the holes through which every single edge passes, and the orders thereof. This comes of course with a decrease in efficiency. It is therefore advised to use the simpler DeltaComplex if this bonus information is not needed. Another drawback is, that the HoleyDeltaComplex structure currently does not work for non-orientable surfaces.

Structures

FlipGraphs.HoleyDeltaComplex — Type

struct HoleyDeltaComplex

An extension of DeltaComplex, keeping track of which edges pass through which holes.

Many operations become slower compared to a DeltaComplex. However this structure is needed, if one wants to compare two different DeltaComplex'.

FlipGraphs.Hole — Type

struct Hole

A structure representing a hole/handle of an orientable surface.

It keeps track of the edges going in and out of the hole.
Although edges are not directed, we give them an arbitrary direction in order to distinct them.
The Hole may be imagined as a line going anticlockwise around the inner surface of the hole.
Edges, crossing this line from right to left are considered to be going into the hole, while edges crossing from left to right are coming out of the hole.

FlipGraphs.Crossing — Type

struct Crossing

A structure representing an edge passing through a hole.

Construction

FlipGraphs.holey_delta_complex — Function

holeyDeltaComplex(g::Integer, num_points::Integer = 1) -> HoleyDeltaComplex

Create a HoleyDeltaComplex of genus g.

FlipGraphs.subdivide! — Method

subdivide!(HD ::HoleyDeltaComplex, t ::Integer)

Add a point to the inside of the t-th TriFace and connect it to each corner.

Common Graph-like methods

As the HoleyDeltaComplex is essentially an extension of the DeltaComplex structure, all methods which take DeltaComplexes also work with HoleyDeltaComplexes.

FlipGraphs.np — Method

np(HD::HoleyDeltaComplex)

Return the number of points in the triangulation defined by HD.

Graphs.nv — Method

nv(HD::HoleyDeltaComplex)

Return the number of vertices(i.e. triangular faces) in HD.

Graphs.ne — Method

ne(HD::HoleyDeltaComplex)

Return the number of edges in the DeltaComplex HD.

This is equal to the number of edges in the triangulation itself.

Graphs.vertices — Method

vertices(HD::HoleyDeltaComplex) -> Vector{TriFace}

Return the list of all vertices in HD.

Graphs.vertices — Method

vertices(HD::HoleyDeltaComplex, d::DualEdge) -> Tuple{TriFace, TriFace}

Return both vertices adjacent to d.

The first TriFace is on the left of the edge, and the second one on the right of the edge.

FlipGraphs.get_vertex — Method

get_vertex(HD::HoleyDeltaComplex, t::Integer) -> TriFace

Graphs.edges — Method

edges(HD::HoleyDeltaComplex) -> Vector{DualEdge}

Return the list of all edges in HD.

Graphs.edges — Method

edges(HD::HoleyDeltaComplex, t::Integer) -> Vector{DualEdge}

Return the list of all 3 edges that are incident to the t-th Triface in HD.

FlipGraphs.get_edge — Method

get_edge(HD::HoleyDeltaComplex, e::Integer) -> DualEdge

FlipGraphs.get_edge — Method

get_edge(HD::HoleyDeltaComplex, t::Integer, side::Integer) -> DualEdge

Return the edge that forms the respective side in the given triangle t.

FlipGraphs.edges_id — Method

edges_id(HD::HoleyDeltaComplex, t::Integer) -> Tuple{Int, Int, Int}

Return the indices of all 3 edges that are incident to the t-th vertex.

FlipGraphs.points — Method

points(HD::HoleyDeltaComplex, d::DualEdge) -> Tuple{Int, Int}

Return both endpoints of d in the triangulation.

FlipGraphs.rename_edges! — Method

rename_edges!(HD::HoleyDeltaComplex, p::Vector{<:Integer})

Rename every vertex(TriFace) in HD, according to the permutation p.

TriFace 1 => TriFace p[1]

FlipGraphs.rename_points! — Method

rename_points!(HD::HoleyDeltaComplex, p::Vector{<:Integer})

Rename all the points in D according to the permutation p.

FlipGraphs.rename_vertices! — Method

rename_vertices!(HD::HoleyDeltaComplex, p::Vector{<:Integer})

Rename every vertex(TriFace) in D, according to the permutation p.

TriFace 1 => TriFace p[1]

Classifying the triangulation

FlipGraphs.euler_characteristic — Method

euler_characteristic(HD::HoleyDeltaComplex) -> Int

Compute the euler characteristic of the HoleyDeltaComplex HD:
$X = #vertices - #edges + #faces$

FlipGraphs.genus — Method

genus(HD::HoleyDeltaComplex) -> Int

Compute the genus of the HoleyDeltaComplex HD.

FlipGraphs.diameter_triangulation — Method

diameter_triangulation(HD::HoleyDeltaComplex) -> Int

Compute the diameter of the triangulation defined by the HoleyDeltaComplex HD.
The diameter of a Graph is the greatest minimal distance between any 2 vertices.

FlipGraphs.diameter_deltaComplex — Method

diameter_deltaComplex(HD::HoleyDeltaComplex)

Compute the diameter of the HoleyDeltaComplex HD.

The diameter of a Graph is the greatest minimal distance between any 2 vertices.

See also diameter_triangulation

FlipGraphs.adjacency_matrix_deltaComplex — Method

adjacency_matrix_deltaComplex(HD::HoleyDeltaComplex) :: Matrix{Int}

Compute the adjacency matrix of the delta complex HD.

FlipGraphs.adjacency_matrix_triangulation — Method

adjacency_matrix_triangulation(HD::HoleyDeltaComplex) :: Matrix{Int}

Compute the simple adjacency matrix of the triangulation defined by HD.

All entries are either 0 or 1.

See also multi_adjacency_matrix_triangulation

FlipGraphs.multi_adjacency_matrix_triangulation — Method

multi_adjacency_matrix_triangulation(HD::HoleyDeltaComplex) :: Matrix{Int}

Compute the adjacency matrix of the multigraph of the triangulation defined by HD.

The $i,j$-th entry not only notes if both points are connected, but also the number of edges that connect these two points.

See also adjacency_matrix_triangulation

FlipGraphs.point_degrees — Method

point_degrees(HD::HoleyDeltaComplex) -> Vector{Int}

Return a vector containing the respective degree of each point in the triangulation.

FlipGraphs.relative_point_degrees — Method

relative_point_degrees(HD::HoleyDeltaComplex, U::Vector{<:Integer}, V::Vector{<:Integer})

Return a vector containing the degree to V for each point in U.

Flipping

FlipGraphs.flip — Method

flip(HD::HoleyDeltaComplex, e::Integer)

Return a copy of HD where the e-th edge has been flipped.

FlipGraphs.flip — Method

flip(HD::HoleyDeltaComplex, d::DualEdge)

Return a copy of HD where the edge d has been flipped.

FlipGraphs.flip! — Method

flip!(HD::HoleyDeltaComplex, e::Integer)

Flip, if possible, the given edge in HD.

FlipGraphs.flip! — Method

flip!(HD::HoleyDeltaComplex, d::DualEdge)

Flip, if possible, the given edge in HD.

FlipGraphs.is_flippable — Method

is_flippable(HD::HoleyDeltaComplex, e::Integer) -> Bool

Return true if the e-th edge can be flipped.

This is always the case if the edge does not connect a TriFace to itself.

FlipGraphs.random_flips! — Method

random_flips!(HD::HoleyDeltaComplex, n::Integer)

Randomly pick an edge, and flip it if possible. Repeat this n times.

HoleyDeltaComplex only

FlipGraphs.num_crossings — Method

num_crossings(H::Hole)

Return the number of Crossing that occur at H.

FlipGraphs.edge_crossings — Method

edge_crossings(HD::HoleyDeltaComplex, e::Integer) -> Vector{Crossing}

Return an array of all the Crossings involving the e-th edge.

The Crossings are ordered along the direction of e.

FlipGraphs.get_crossing — Method

get_crossing(H::Hole) -> Crossing

Return any Crossing of the hole H.

Usually the Crossing which is returned, was the last one added to the Hole.

FlipGraphs.remove_holeloops! — Method

remove_holeloops!(HD::HoleyDeltaComplex)

Remove all pairs of crossings where an edge enters a hole, and imeadiately comes out again.