These are some functions, which are in general not usefull to a person using the package:

FlipGraphs.FGVertexCandidate — Type

struct FGVertexCandidate

A candidate for a possible vertex in a flip graph.

It hasn't yet been decided if this candidate is a new vertex in the graph, or isomorph to an already existing DeltaComplex.

FlipGraphs.is_isomorphic — Method

is_isomorphic(candidate::FGVertexCandidate, fgv::FGVertex; kwargs...) -> Bool

Return true if candidate is in the isotopy class of fgv.

Arguments

labeled_points :: Bool = true : If is set to false, then the isomorphism would also allow a relabeling of the points.

FlipGraphs.rename_vertices! — Function

rename_vertices!(g::TriangulatedPolygon, p::Vector{<:Integer}) :: TriangulatedPolygon

Rename the vertices of the triangulated convex polygon g by applying the permutation p.

rename_vertices!(D::DeltaComplex, p::Vector{<:Integer})

Relabel every vertex(TriFace) in D, according to the permutation p.

TriFace 1 => TriFace $p[1]$