Triangulations of Convex Polygons

In order to better understand triangulations, flips and flipgraphs, it can be helpfull to start simple. If you take any number of points, and iteratively connect them with staight edges that do not cross each other, until you can no longer add an edge that does not cross any other edge, what you'll get is a triangulation.

To get from one triangulation of these points to another, you may choose an inner edge, and flip it. If you look at any inner edge, the two triangles adjacent to it form a quadrilateral with the edge as one of its diagonals. By flipping an edge, all we have to do is replace it by the other diagonal. As we don't want to have straight edges that cross each other, a flip can only be done, if the quadrilateral is convex, and no three of its corners lie on the same line. As we are interested in taking this theory to closed surfaces, where we will no longer restrict the edges to being straight, we will only consider triangulations of points in convex general position. In this case, a triangulation in the geometric sense is equivalent to a triangulation in the graph sense.

In graph theory, we do not care, if edges are straight or curved. In this case, a triangulation is simply a maximal(in regards to the number of edges) planar graph with a fixed number of points. Here, we do not care where the points are located, however, in order keep in line with the geometric sense and have a simple visualization,we will consider these points to be the vertices of a convex polygon (i.e. points in convex position).

FlipGraphs.TriangulatedPolygon — Type

struct TriangulatedPolygon <: AbstractGraph{Integer}

A structure representing a triangulation of a convex polygon.

TriangulatedPolygon implements the AbstractGraph interface from Graphs.jl. It is therefore possible to use it with other packages that work with Graphs.jl. This is very helpfull for plotting the graph.\ Vertices are not explicity stored in TriangulatedPolygon. Only the total number of vertices is stored. They are implicitely labeled by the integers from 1 up to the total number of vertices.

Edges are stored as an adjacency list of which vertices are connected to another.

Constructors

FlipGraphs.triangulated_polygon — Function

triangulated_polygon(n::Integer) -> TriangulatedPolygon

Create a triangulated convex n-gon.

Vertices are named from 1 to n in an anticlockwise manner. The inside is triangulated in a zig-zag manner.

Graph Methods

Graphs.nv — Method

nv(g::TriangulatedPolygon)

Return the number of vertices in g.

Graphs.ne — Method

ne(g::TriangulatedPolygon)

Return the number of edges in g.

Graphs.vertices — Method

vertices(g::TriangulatedPolygon)

Return a list of all the vertices in g.

Graphs.edges — Method

edges(g::TriangulatedPolygon) -> Vector{Edges}

Compute and return a list of all the edges in g.

Edges are not directed. It is however necessary to define a source and a target. For TriangulatedPolygon, the source will be the incident vertex with the smaller id.

Graphs.has_vertex — Method

ne(g::TriangulatedPolygon)

Return the number of edges in g.

Graphs.has_edge — Method

has_edge(g::TriangulatedPolygon, e::Edge)

Graphs.has_edge — Method

has_edge(g::TriangulatedPolygon, s, d)

Graphs.neighbors — Method

neighbors(g::TriangulatedPolygon, v::Integer)

Return a list of all the vertices in g that are adjacent to v.

FlipGraphs.degrees — Method

degrees(g::TriangulatedPolygon) -> Vector{Int}

Return a list of the degrees of every single vertex in g

FlipGraphs.is_isomorph — Method

is_isomorph(g1::TriangulatedPolygon, g2::TriangulatedPolygon, permutations::Vector{Vector{Integer}})

Return true if g1 is identical to g2 up to a renaming of the vertices of g1 by one of the given permutations.

FlipGraphs.rename_vertices — Method

rename_vertices(g::TriangulatedPolygon, p::Vector{<:Integer})

Rename the vertices of g by applying the permutation p.

FlipGraphs.adjacency_matrix — Method

adjacency_matrix(g::TriangulatedPolygon)

Compute the adjacency matrix for the triangulated graph g.

Flipping

FlipGraphs.is_flippable — Method

is_flippable(g::TriangulatedPolygon, src::Integer, dst::Integer) -> Bool

Return whether or not the edge can be flipped.

Note that for a triangulation of a convex polygon inner edges are always flippable, while outer edges cannot be flipped.

FlipGraphs.is_flippable — Method

is_flippable(g::TriangulatedPolygon, e::Edge) -> Bool

FlipGraphs.flip! — Method

flip!(g::TriangulatedPolygon, src::Integer, dst::Integer)

Flip the the edge incident to src and dst in g.

FlipGraphs.flip! — Method

flip!(g::TriangulatedPolygon, e::Edge)

Flip e in g.

FlipGraphs.flip — Method

flip(g::TriangulatedPolygon, e::Edge)

Return the triangulated polygon obtained by flipping the edge e in g.