FlipGraphs.jl Documentation

Welcome to the documentation of the FlipGraphs package. This package is part of my master's thesis on triangulations of closed surfaces.

The aim of the package, is to model triangulations on any closed surface and to compute their respective flipgraphs.

FlipGraphs — Module

FlipGraphs

A package for triangulations of closed surfaces and their respective FlipGraphs.