DeltaComplexes

Modelling a triangulation on a closed surface

A Δ-complex is a representation of a triangulation on a closed surface. To define a triangulation on a closed surface, it does not suffice to take into account vertices and edges. We will also need to take into account the triangular faces between them. Therefore the triangulations are modeled using an extension of their dual graph.

Vertices are triangular faces which in turn consist of three points, and three edges. These points and edges are not necessarily distinct. Edges in the dual Graph(i.e. the Δ-complex) connect two triangular faces if they in turn share an edge. In order to avoid confusion between the edges of the triangulation and the edges in the dual graph, I will, hence forwards call the latter dual edge.

FlipGraphs.DeltaComplex — Type

struct DeltaComplex

A Graph datastructure representing a triangulation of a surface.

The DeltaComplex may be thought of as the dual of a triangulation.

Vertices are triangular faces (TriFace). Every vertex has three edges (DualEdge) incident to it.

FlipGraphs.DualEdge — Type

struct DualEdge

Representation of an edge in a DeltaComplex (i.e. the dual graph of a triangulation).

A DualEdge connects two TriFaces through specific sides.

FlipGraphs.TriFace — Type

struct TriFace

A TriFace represents a triangle in the triangulization of a surface.

The TriFace's form the vertices of a DeltaComplex. Each TriFace is formed between 3 points and is connected to 3 Trifaces through DualEdges. The points and neighboring triangles do not have to be unique.

The points and edges are stored in an anticlockwise order.
The first edge/side is between the first and second point.
The second edge/side is between the second and third point.
The third edge/side is between the third and first point.

Construction of a DeltaComplex

This Module comes with some handy and easy to use tools to construct a triangulation of a surface:

FlipGraphs.delta_complex — Function

delta_complex(genus , num_points = 1)

Create a triangulation of an orientable surface with num_points points on it.

By default num_points is set to 1.

delta_complex(s :: Array{<:Integer,1})

Create a triangulation of an orientable surface with a single point, by gluing corresponding edges together.
s should be an array of nonzero integers representing the edges of a polygon in anticlockwise order.\ The i-th edge is orientated anticlockwise if s[i]>0 and anticlockwise if s[i]<0.
If s[i] and s[j] have the same absolute value, they are glued together while respecting their orientation.
#Examples The following results in the triangulation of a torus with one point:

julia> delta_complex([1,2,-1,-2])

The following results in the triangulation of a Klein bottle with one point:

julia> delta_complex([1,2,-1,-2])

FlipGraphs.subdivide! — Method

subdivide!(D::DeltaComplex, t::Integer)

Add a point to the inside of the t-th TriFace and connect it to each corner.

Common Graph-like methods

DeltaComplex is not an implementation of Graphs.AbstractGraph. However, as it is similar to a simple graph, I used the same notation and function names for simplicity

FlipGraphs.np — Method

np(D::DeltaComplex)

Return the number of points in the triangulation defined by D.

Graphs.nv — Method

nv(D::DeltaComplex)

Return the number of vertices(i.e. triangular faces) in D.

Graphs.ne — Method

ne(D::DeltaComplex)

Return the number of edges in the DeltaComplex D.

This is equal to the number of edges in the triangulation itself.

Graphs.vertices — Method

vertices(D::DeltaComplex) -> Vector{TriFace}

Return the list of all vertices in D.

FlipGraphs.get_vertex — Method

get_vertex(D::DeltaComplex, t::Integer) -> TriFace

FlipGraphs.get_vertex_id — Method

get_vertex_id(d::DualEdge, side::Integer) -> Int

Return the index of the vertex (TriFace) to the left ofd`.

Graphs.edges — Method

edges(D::DeltaComplex) -> Vector{DualEdge}

Return the list of all the DualEdges in D.

FlipGraphs.get_edge — Method

get_edge(D::DeltaComplex, e::Integer) -> DualEdge

FlipGraphs.get_edge — Method

get_edge(D::DeltaComplex, t::Integer, side::Integer) -> DualEdge

Return the edge that forms the respective side in the given triangle t.

FlipGraphs.id — Method

id(d::DualEdge) -> Int

Return the index of d in its DeltaComplex.

FlipGraphs.points — Method

points(T::TriFace) -> Tuple{Int, Int, Int}

Return a tuple of the three points, that form the corners of T.

FlipGraphs.get_point — Method

get_point(T::TriFace, corner::Integer) -> Int

FlipGraphs.points — Method

points(D::DeltaComplex, d::DualEdge) -> Tuple{Int, Int}

Return both endpoints of d in the triangulation.

Graphs.vertices — Method

vertices(D::DeltaComplex, d::DualEdge) -> Tuple{TriFace, TriFace}

Return both vertices adjacent to d.

The first TriFace is on the left of the edge, and the second one on the right of the edge.

FlipGraphs.vertices_id — Method

vertices_id(d::DualEdge) -> Tuple{Int, Int}

Return the indices of both vertices (TriFaces) adjacent to d.

FlipGraphs.get_edge — Method

get_edge(T::TriFace, side::Integer) -> DualEdge

FlipGraphs.get_edge_id — Method

get_edge_id(T::TriFace, side::Integer) -> Integer

Return the index of the edge on the respective side in T.

Graphs.edges — Method

edges(T::TriFace) -> Vector{DualEdge}

Return the list of all 3 edges that are incident to T.

Graphs.edges — Method

edges(D::DeltaComplex, t::Integer) -> Vector{DualEdge}

Return the list of all 3 edges that are incident to the t-th Triface in D.

FlipGraphs.edges_id — Method

edges_id(T::TriFace) -> Tuple{Int, Int, Int}

Return the indices of all 3 edges that are incident to T.

FlipGraphs.id — Method

id(T::TriFace)

Return the index of T in its DeltaComplex.

FlipGraphs.other_endpoint — Method

other_endpoint(d::DualEdge, t::Integer, side::Integer) -> Tuple{Int, Int8}

Return the index of the other TriFace and its respective side, that is incident to d

FlipGraphs.is_similar — Method

is_similar(d1::DualEdge, d2::DualEdge) -> Bool

Return true if d1 and d2 have the same twist and are connected to the same triangles.

This is only the case if d1 and d2 are the same edge, or if they are incident to a point of degree 2.

Classifying the triangulation

Here are some usefull methods, to pull out general information about the Δ-Complex, and the triangulation it represents:

FlipGraphs.euler_characteristic — Method

euler_characteristic(D::DeltaComplex)

Compute the euler characteristic of the DeltaComplex D:
$X = #vertices - #edges + #faces$

FlipGraphs.genus — Method

genus(D::DeltaComplex)

Compute the genus of the DeltaComplex D if it forms an orientable surface.

FlipGraphs.diameter_triangulation — Method

diameter_triangulation(D::DeltaComplex)

Compute the diameter of the triangulation defined by the DeltaComplex D.

The diameter of a Graph is the greatest minimal distance between any 2 vertices.

See also diameter_deltaComplex

FlipGraphs.diameter_deltaComplex — Method

diameter_deltaComplex(D::DeltaComplex)

Compute the diameter of the DeltaComplex D.
The diameter of a Graph is the greatest minimal distance between any 2 vertices.

See also diameter_triangulation

FlipGraphs.adjacency_matrix_deltaComplex — Method

adjacency_matrix_deltaComplex(D::DeltaComplex) :: Matrix{<:Integer}

Compute the adjacency matrix of the delta complex D.

FlipGraphs.adjacency_matrix_triangulation — Method

adjacency_matrix_triangulation(D::DeltaComplex) :: Matrix{Int}

Compute the simple adjacency matrix of the triangulation defined by D.

All entries are either 0 or 1.

See also multi_adjacency_matrix_triangulation

FlipGraphs.multi_adjacency_matrix_triangulation — Method

multi_adjacency_matrix_triangulation(D::DeltaComplex) :: Matrix{Int}

Compute the adjacency matrix of the multigraph of the triangulation defined by D.

The i,j-th entry not only notes if both points are connected, but also the number of edges that connect these two points.

See also adjacency_matrix_triangulation

FlipGraphs.point_degrees — Method

point_degrees(D::DeltaComplex) -> Vector{<:Integer}

Return a vector containing the respective degree of each point in the triangulation.

FlipGraphs.relative_point_degrees — Method

relative_point_degrees(D::DeltaComplex, U::Vector{<:Integer}, V::Vector{<:Integer})

Return a vector containing the degree to V for each point in U.

FlipGraphs.rename_edges! — Method

rename_edges!(D::DeltaComplex, p::Vector{<:Integer})

Rename every vertex(TriFace) in D, according to the permutation p.

TriFace 1 => TriFace p[1]

FlipGraphs.rename_points! — Method

rename_points!(D::DeltaComplex, p::Vector{<:Integer})

Rename all the points in D according to the permutation p.

FlipGraphs.rename_vertices! — Method

rename_vertices!(D::DeltaComplex, p::Vector{<:Integer})

Rename every vertex(TriFace) in D, according to the permutation p.

TriFace 1 => TriFace p[1]

Non-orientable closed surfaces

Regarding on which way edges are glued together, the resulting surface may be non-orientable. These surfaces can also be modeled by DeltaComplex and all of the methods above (with the exception of genus) may still be applied.

FlipGraphs.is_orientable — Method

is_orientable(D::DeltaComplex)

Ckeck if the surface on which D lies is orientable or not.

FlipGraphs.demigenus — Function

demigenus(D::DeltaComplex)

Compute the demigenus of the DeltaComplex D.
The demigenus or non-orientable genus ($k$) of a connected non-orientable closed surface is defined via the euler characteristic ()$X$) :
$k = 2 - X$

FlipGraphs.twist_edges! — Function

twist_edges!(D::DeltaComplex, t::Integer)
twist_edges!(T::TriFace)

Twist or untwist all 3 edges in a TriFace, and reverse the side order.

This action gives an equivalent representation ot the same triangulation. It is usefull in the case that you would like to untwist a certain edge.

Examples

julia> D = delta_complex([1,2,-1,2]);
julia> T = get_vertex(D,1);
julia> edges(T)
3-element Array{DualEdge,1}:
 DualEdge 2 : (Δ1)-(1)-------(2)-(Δ2)
 DualEdge 1 : (Δ1)-(2)-------(3)-(Δ2)
 DualEdge 3 : (Δ2)-(1)---↺---(3)-(Δ1)
 julia> twist_edges!(T);
 julia> edges(T)
 3-element Array{DualEdge,1}:
 DualEdge 3 : (Δ2)-(1)-------(1)-(Δ1)
 DualEdge 1 : (Δ1)-(2)---↺---(3)-(Δ2)
 DualEdge 2 : (Δ1)-(3)---↺---(2)-(Δ2)

Flipping

A flip is defined as the action of replacing an edge in the triangulation by the other diagonal of the quadrilateral formed by the two triangles adjacent to the edge.

It has been shown, that the flipgraph of any closed surface is connected, hence it is possible, to obtain any triangulation by a finite number of flips.

FlipGraphs.flip! — Method

flip!(D::DeltaComplex, e::Integer)

Flip, if possible, the given edge in D.

FlipGraphs.flip! — Method

flip!(D::DeltaComplex, d::DualEdge)

Flip, if possible, the given edge in D.

FlipGraphs.is_flippable — Method

is_flippable(D::DeltaComplex, e::Integer)

Return true if the given edge can be flipped.

This is always the case if the edge does not connect a TriFace to itself.

FlipGraphs.is_flippable — Method

is_flippable(d::DualEdge)

Return true if the given edge is can be flipped.

This is always the case if the edge does not connect a TriFace to itself.

FlipGraphs.random_flips! — Method

random_flips!(D::DeltaComplex, n::Integer)

Randomly pick an edge, and flip it if possible. Repeat this n times.

FlipGraphs.randomize! — Method

randomize!(D::DeltaComplex; kwargs...) -> Int

Randomly flip edges in D until D is sufficiently generic.

Return the number of attempted flips.

The measure by which we determin if D is sufficiently generic is through its diameter. This Method repeatedly flips a certain number of times. After each flip sequence the diameter is computed. Once this was repeated a certain number of times, the variance of all these past diameter measurements gets computed.

In theory, the variance should diminish over time. However, as we are flipping randomly, it will never truly converge to 0. A certain flutter in the variance is expected, this will at some point cause the variance to increase every so often. The algorithm stops once the last measured variance is bigger than the past few measurements.

Arguments

num_initial_flips::Integer=1000000 : the number of flips to do before even start taking measurements.

num_flips_per_step::Integer=10000 : the number of flips to do before computing the diameter each step.

variance_interval_size::Integer=10 : the number of diameters to store, before computing their variance.

lookback_size::Integer=2 : how far back to compare the current variance to, in order to decide when to stop.

Examples

julia> D = delta_complex(30,30);
julia> randomize!(D, num_initial_flips=10000, num_flips_per_step = 1000, variance_interval_size=10, lookback_size = 5)
160000