General utilities

These are some helping methods which youmight want to use, but aren't specific to one of the main structures in this package.

Graphs.diameter — Method

diameter(adjacency_matrix :: Matrix{<:Integer}) :: Int

Compute the diameter of a graph from its simple adjacency matrix.

All values in adjacency_matrix should be either 0 or 1.

FlipGraphs.adjacency_matrix — Method

adjacency_matrix(adjList::Vector{Vector{<:Integer}}) -> Matrix{Int}

Construct the adjacency matrix from an adjaceny list.

FlipGraphs.invert_permutation — Method

invert_permutation(p::Vector{<:Integer})

Return the inverse of the permutation p.

Example

julia> p = [2,1,4,5,3];
julia> p_inv = invert_perm(p); 
julia> show(p_inv)
[4, 2, 5, 1, 6, 3]
julia> show(p_inv[p])
[1, 2, 3, 4, 5, 6]

FlipGraphs.degrees — Method

degrees(A::Matrix{<:Integer}) -> Vector{<:Integer}

Return a vector containing the degrees of every vertex given an adjacency matrix A.